requirement.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-bootstrap4==3.0.0
django-ckeditor==6.0.0
django-filter==2.4.0
django-import-export==2.5.0
django-js-asset==1.2.2
djangorestframework==3.12.4
et-xmlfile==1.0.1
html5lib==1.1
idna==2.10
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
PyJWT==1.7.1
PyPDF2==1.26.0
pytube==10.7.2
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
twilio==6.56.0
urllib3==1.26.4
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
